check if user exists

+++ bootcamp/create-user.sh
@@ -2,7 +2,9 @@
 set -e
 
 # delete user
-sudo userdel -rf bootcamp
+user_exists(){ id "$1" &>/dev/null; } # silent, it just sets the exit code
+if user_exists "$1"; code=$?; then  # use the function, save the code
+    sudo userdel -rf bootcamp
 
 # creates a standerd user
 sudo adduser --disabled-password --gecos "" bootcamp
